Restructure directory tree
