+++ comm-web/src/components/banner/res/github.png
